+++ src/data/PollOptionModel.mjs
@@ -1,6 +1,8 @@
 const {StringField, NumberField} = foundry.data.fields;
 
 export default class PollOptionModel extends foundry.abstract.DataModel {
+  static maxLength = 30;
+
   static defineSchema() {
     return {
       label: new StringField(),
@@ -9,7 +11,19 @@ export default class PollOptionModel extends foundry.abstract.DataModel {
     };
   }
 
+  get tooltip() {
+    return `<h4>${this.label}</h4><p>${this.votersTooltip}</p>`;
+  }
+
   get votersTooltip() {
-    return this.voters ? Object.values(this.voters).join(", ") : "";
+    return (game.user.isGM || this.parent.settings.secret === false) && this.voters
+      ? Object.values(this.voters).join(", ")
+      : "";
+  }
+
+  get labelLimited() {
+    return this.label.length > PollOptionModel.maxLength
+      ? `${this.label.slice(0, PollOptionModel.maxLength)}...`
+      : this.label;
   }
 }